Rename core to common, remove re-exports from common
