+++ components/common/ExitPreviewButton.tsx
