+++ components/common/Hero/Hero.tsx
