+++ components/common/Hero/index.ts
@@ -1,0 +1,1 @@
+export { default as Hero } from './Hero'
