+++ components/common/Layout/Layout.tsx
@@ -1,6 +1,6 @@
-import Footer from '../Footer/Footer'
-import Header from '../Header/Header'
-import Nav from '../Nav/Nav'
+import { Footer } from '../Footer'
+import { Header } from '../Header'
+import { Nav } from '../Nav'
 import OfflineBanner from '../OfflineBanner'
 
 type Props = {
