+++ components/common/Markdown/index.ts
@@ -1,0 +1,1 @@
+export { default as Markdown } from './Markdown'
