+++ components/common/OfflineBanner.tsx
