--- components/core/index.ts
@@ -1,1 +1,0 @@
-export { default as Layout } from './Layout/Layout'
