+++ pages/contributors/index.tsx
@@ -1,7 +1,7 @@
 import Contributor from '@components/contributor/Contributor'
 import ContributorFeatured from '@components/contributor/ContributorFeatured'
-import { Layout } from '@components/core'
-import Hero from '@components/core/Hero/Hero'
+import { Layout } from '@components/common/Layout'
+import Hero from '@components/common/Hero/Hero'
 import { fetchAPI } from '@lib/api'
 import { partition } from '@lib/partition'
 import { InferGetStaticPropsType } from 'next'
